Adding npm updates and relocation of a component
