+++ package.json
@@ -4,6 +4,8 @@
   "private": true,
   "dependencies": {
     "@microsoft/applicationinsights-react-js": "^3.0.0",
+    "css-select": "^5.1.0",
+    "nth-check": "^2.1.1",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
